requirements.txt
pandas
polars
pyarrow
requests
yfinance
nltk
matplotlib
seaborn
scikit-learn
notebook
jupyter
atproto
python-dotenv
